all functions moved to file "functions.py" and added tests to file "main.py"
